20190613-heartbeat-v2: Review fixes 3
